@@ -13,7 +13,7 @@
     "psr-4": { "Knojector\\SteamAuthenticationBundle\\": "" }
   },
   "require": {
-    "php": ">=7.1.3",
+    "php": ">=8.0.0",
     "symfony/config": "^4.0|^5.0",
     "symfony/dependency-injection": "^4.0|^5.0",
     "symfony/event-dispatcher": "^4.0|^5.0",
