@@ -14,15 +14,8 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
  */
 class SteamCallbackResolver implements ArgumentValueResolverInterface
 {
-    /**
-     * @var ValidatorInterface
-     */
-    private $validator;
-
-    public function __construct(ValidatorInterface $validator)
-    {
-        $this->validator = $validator;
-    }
+    public function __construct(private ValidatorInterface $validator)
+    {}
 
     /**
      * @inheritDoc
