@@ -9,15 +9,8 @@ use Symfony\Contracts\EventDispatcher\Event;
  */
 abstract class CommunityIdAwareEvent extends Event
 {
-    /**
-     * @var string
-     */
-    private $communityId;
-
-    public function __construct(string $communityId)
-    {
-        $this->communityId = $communityId;
-    }
+    public function __construct(private string $communityId)
+    {}
 
     public function getCommunityId(): string
     {
